remove s3 service
