+++ app-server/src/api/v1/datasets.rs
@@ -15,7 +15,7 @@ use crate::{
     query_engine::QueryEngine,
     routes::{PaginatedResponse, types::ResponseResult},
     sql::{self, ClickhouseReadonlyClient},
-    storage::StorageService,
+    storage::{Storage, StorageTrait},
 };
 
 #[derive(Deserialize)]
@@ -324,7 +324,7 @@ async fn create_datapoints(
 async fn get_parquet(
     path: web::Path<(String, String)>,
     db: web::Data<DB>,
-    storage: web::Data<Arc<StorageService>>,
+    storage: web::Data<Arc<Storage>>,
     project_api_key: ProjectApiKey,
 ) -> ResponseResult {
     let (dataset_id_str, name) = path.into_inner();
@@ -348,11 +348,9 @@ async fn get_parquet(
             "error": "exports storage is not configured"
         })));
     };
-    let content_length = storage.get_size(project_id, &bucket, &parquet_path).await?;
+    let content_length = storage.get_size(&bucket, &parquet_path).await?;
 
-    let get_response = storage
-        .get_stream(project_id, &bucket, &parquet_path)
-        .await?;
+    let get_response = storage.get_stream(&bucket, &parquet_path).await?;
 
     let filename = parquet_path.split('/').last().unwrap_or(&name);
