+++ app-server/src/routes/mod.rs
@@ -1,5 +1,4 @@
 pub mod error;
-pub mod payloads;
 pub mod probes;
 pub mod realtime;
 pub mod rollouts;
